+++ Source/Private/ASCollectionLayout.mm
@@ -93,7 +93,7 @@ static const ASScrollDirection kASStaticScrollDirection = (ASScrollDirectionRigh
   ASCollectionLayoutState *layout = [context.layoutDelegateClass calculateLayoutWithContext:context];
   [context.layoutCache setLayout:layout forContext:context];
 
-  // Measure elements in the measure range ahead of time, block on the initial rect as it'll be visible shortly
+  // Measure elements in the measure range ahead of time
   CGSize viewportSize = context.viewportSize;
   CGPoint contentOffset = context.initialContentOffset;
   CGRect initialRect = CGRectMake(contentOffset.x, contentOffset.y, viewportSize.width, viewportSize.height);
@@ -101,7 +101,11 @@ static const ASScrollDirection kASStaticScrollDirection = (ASScrollDirectionRigh
                                                                    kASDefaultMeasureRangeTuningParameters,
                                                                    context.scrollableDirections,
                                                                    kASStaticScrollDirection);
-  [self _measureElementsInRect:measureRect blockingRect:initialRect layout:layout];
+  // The first call to -layoutAttributesForElementsInRect: will be with a rect that is way bigger than initialRect here.
+  // If we only block on initialRect, a few elements that are outside of initialRect but inside measureRect
+  // may not be available by the time -layoutAttributesForElementsInRect: is called.
+  // Since this method is usually run off main, let's spawn more threads to measure and block on all elements in measureRect.
+  [self _measureElementsInRect:measureRect blockingRect:measureRect layout:layout];
 
   return layout;
 }
@@ -248,11 +252,7 @@ static const ASScrollDirection kASStaticScrollDirection = (ASScrollDirectionRigh
   }
 
   // Step 2: Get layout attributes of all elements within the specified outer rect
-  ASCollectionLayoutContext *context = layout.context;
-  CGSize pageSize = context.viewportSize;
-  ASPageToLayoutAttributesTable *attrsTable = [layout getAndRemoveUnmeasuredLayoutAttributesPageTableInRect:rect
-                                                                                                contentSize:contentSize
-                                                                                                   pageSize:pageSize];
+  ASPageToLayoutAttributesTable *attrsTable = [layout getAndRemoveUnmeasuredLayoutAttributesPageTableInRect:rect];
   if (attrsTable.count == 0) {
     // No elements in this rect! Bail early
     return;
@@ -260,6 +260,8 @@ static const ASScrollDirection kASStaticScrollDirection = (ASScrollDirectionRigh
 
   // Step 3: Split all those attributes into blocking and non-blocking buckets
   // Use ordered sets here because some items may span multiple pages, and the sets will be accessed by indexes later on.
+  ASCollectionLayoutContext *context = layout.context;
+  CGSize pageSize = context.viewportSize;
   NSMutableOrderedSet<UICollectionViewLayoutAttributes *> *blockingAttrs = hasBlockingRect ? [NSMutableOrderedSet orderedSet] : nil;
   NSMutableOrderedSet<UICollectionViewLayoutAttributes *> *nonBlockingAttrs = [NSMutableOrderedSet orderedSet];
   for (id pagePtr in attrsTable) {